修改了ModbusMasterTransaction中关于是否有数据要读的判断，改为用available判断而不是read()的返回值是否是-1

+++ src/Arduino.h
@@ -38,7 +38,7 @@ public:
 			char stopbit);
 	~Stream(void);
 	void flush(void);
-	bool begin(uint16_t baudrate);
+	bool begin(int baudrate);
 	void close(void);
 	DWORD write(uint8_t data);
 	uint8_t read(void);
